refactor: 將getUser移到middleware
