+++ middleware/general-data.js
@@ -1,0 +1,54 @@
+const { User, Followship, Tweet } = require('../models')
+const helpers = require('../_helpers')
+
+const getUser = async (req, res, next) => {
+  // 取得loginUser(使用helpers), userId
+  const loginUser = helpers.getUser(req)
+  const { userId } = req.params
+  try {
+    // 判斷session儲存的資料是否跟req相同
+    if (req.session.userData?.id === userId) return next()
+    // 取對應的user資料
+    const user = await User.findByPk(userId)
+    // 判斷user是否存在，沒有就err
+    if (!user) throw new Error('該用戶不存在!')
+    // 變數存，user是否為使用者
+    const isLoginUser = user.id === loginUser.id
+
+    // 取得following跟follower的count
+    const [FollowingsCount, FollowersCount, tweetsCount] = await Promise.all([
+      // 計算user的folowing數量
+      Followship.count({
+        where: { followerId: userId }
+      }),
+      // 計算user的folower數量
+      Followship.count({
+        where: { followingId: userId }
+      }),
+      // 推文及推文數
+      Tweet.count({
+        where: { UserId: userId }
+      })
+    ])
+
+    // 將變數加入session
+    req.session.userData = {
+      ...user.toJSON(),
+      cover: user.cover || '/images/profile/cover.png',
+      avatar: user.avatar || 'https://ionicframework.com/docs/img/demos/avatar.svg',
+      FollowingsCount,
+      FollowersCount,
+      tweetsCount,
+      isLoginUser,
+      isFollowing: isLoginUser ? null : loginUser.Followings.some(following => following.id === user.id)
+    }
+    // next
+    return next()
+  } catch (err) {
+    next(err)
+  }
+}
+
+module.exports = {
+  getUser
+}
